Decouple SSE servlets from the streams that produce JSON to allow alternate methods of consumption
